feat: Make more-like-this (MLT) more usable (#3328)
# Ticket(s) Closed

- Closes #

## What

Docs, various fixing, and testing for the various
behaviors/configurations of MLT.

- By default, the MLT query was silently excluding terms that appeared
less than twice in the input doc, or in fewer than 5 documents in the
entire index. These defaults have been removed so that no terms are
silently excluded by default.
- By default the MLT query would match against every field in the index,
but there was no way to specify a list of fields. Elastic has this
[ability](https://www.elastic.co/docs/reference/query-languages/query-dsl/query-dsl-mlt-query)
and now so do we.
- The various configuration options of MLT were entirely untested and
undocumented, now they have been.
- Tantivy's MLT query does not support JSON fields or JSON paths -- we
now have an error/docs for this.
- If a `document_id` is provided that does not exist in the heap, we
return an empty result set instead of erroring.
- The "stopwords" argument was snake case `stop_words` which I didn't
like because our stopwords filter is just one word. Now it's just
`stopwords`.

## Why

Better UX

## How

## Tests

Added a regression test