ci: Fix broken checkouts on publish workflows (#3385)
# Ticket(s) Closed

- Closes #N/A

## What
The checkout fails when using the ParadeDB bot token, and its not
necessary for the checkout anyways.

Upgrading to checkoutv5 seems to require git to be explicitly installed
before, so I’ve also done that. This needs to be tested in a dry run
before merging, but should fix the enterprise binaries not publishing.

I’ve also added Slack alerts so we know if the deploy fails earlier in
the future.
## Why

## How

## Tests