ci: Also test on arm (#3407)
# Ticket(s) Closed

- Closes #N/A

## What
Rigth now our CI only tests on AMD64. This adds ARM64 to all of our
`pg_search` tests. It does not add it to our stressgres/benchmarks,
though. That can come next if we feel the need for it.

## Why
One of our community users reported segfaults running `index_layer_size`
and it seems the main difference between his dev and prod environments
is that prod is on arm.

## How
^

## Tests
CI