perf: Optimize `TermSet` for very large sets of terms. (#3412)
Further optimizes `pdb.term_set` and `paradedb.term_set`, and deprecates
using `pdb.term_set` in aggregate position.

switching to using fast fields when more than `1024` terms were used in
the set. But there was more that could be done.

For a `paradedb.aggregate` query using a `pdb.term_set` constructed from
an `array_agg` containing 10mm distinct inputs and 8 segments, this PR
further optimizes the fast field execution path:

| version | runtime |
| ------- | -------- |
| pre-#3351 - 0 workers | 35.851 s |
| pre-#3351 - 8 workers | swapping - did not complete |
| #3351 - 0 workers | 12.573 s |
| #3351 - 8 workers | 13.708 s |
| #3412 - 0 workers | 5.532 s |
| #3412 - 8 workers | 8.538 s |

Before #3351, the posting-list based execution mode for term sets was
not able to complete on my machine with multiple workers, because it
required enough memory to trigger swapping.

Critical to note: in the case of a massive `pdb.term_set` like this,
additional workers might be a pessimization. That's because the cost of
propagating and creating the query is expensive enough that it can dwarf
the actual aggregate time. For larger segment counts or larger
aggregates, the results might be different.

Additionally, this change deprecates using `pdb.term_set` in aggregate
position (as added in #3336): using `pdb.term_set` in function position
with an `array_agg` is equivalent, and sometimes slightly faster.

* Incorporates https://github.com/paradedb/tantivy/pull/75
* Removes some allocation from `pdb.term_set` and `paradedb.term_set`
creation.
* Skips allocation `FieldName::path` arrays if they contain a single
component.