fix: FSM double-add/undefined behavior issues (#3426)
# Ticket(s) Closed

- Closes #

## What

I have identified two places where we add the same blocks to the FSM,
which leads to undefined behavior.

1. In `garbage_collect_index`, we go through the segment meta entries
list and return all blocks belonging to files of recyclable meta entries
to the FSM. However, there's nothing guarding against two transactions
from seeing the same recyclable entries twice.

2. Inside `drain` of the V2FSM, when the "head" block of an AVL leaf is
drained, we unlink it and send it to the FSM. However, it is possible
for two concurrent transactions to both think they've drained the head,
and send the same head block back to the FSM.

The problem arises here:

```rust
if !modified {
    // we didn't change anything
    buffer.set_dirty(false);
}

// drop the leaf buffer -- we're done with it and it's possible we'll need to
// unlink it from the list and that requires an exclusive lock on the tree
// and we can't have both at the same time
drop(buffer);

if should_unlink_head {
    let old_head = head_blockno;

    // get mutable tree without holding any other locks
    let mut root = bman.get_buffer_mut(self.start_blockno);
```

Suppose transaction A drains the "head" block. Before unlinking the
head, it must drop it (see `drop(buffer)`). When the buffer is dropped,
a concurrent transaction B can lock onto the same "head" block, see that
it's empty, and think that it is the one that drained the head block.
Now both transactions go to unlink the same block and return it to the
FSM.

## Why

## How

To fix #1, when recycling meta entries we check to see if the entry has
already been deleted using Postgres' `LP_DEAD` flag.

To fix #2, before sending the "head" block to the FSM we re-check the
root to see if the head block is still what we think it is. If another
transaction has already unlinked the head then the values won't match.

## Tests

Ran stressgres `single-server.toml` for a long time with `block_tracker`
enabled.