docs: Significant docs overhaul for SEO (#3422)
# Ticket(s) Closed

- Closes #N/A

## What
This PR improves our docs' SEO on many counts:

- Fixes an issue where some pages had the same title
- Fixes an issue where some pages' canonical tags were incorrect
- Removes an incorrect blog/ redirect
- Fixed an issue in parsing the `localhost` mention in Kubernetes
- Add missing canonical tags to all the pages that were missing them
(mostly in /v2/)
- Remove underscores in URLs and replace them with dashes
- Add missing description to all docs pages
- Add a check in our `test-docs.yml` CI to enforce these settings for
all docs pages, to avoid breaking this in the future (especially on new
pages)

Don't be surprised — mainly because of filepath renames, the diff is
massive.

## Why
Proper SEO is important for users finding our docs.

## How
Follow Semrush

## Tests
Local and CI

---------

Co-authored-by: Ming Ying <ming.ying.nyc@gmail.com>