perf: Avoid copying Postgres buffers (#3402)
# Ticket(s) Closed

- Closes #3374

## What

This change introduces zero-copy from Postgres `Buffer`s (in most
cases), meaning that we avoid copying out of buffers in order to hand
out `OwnedBytes` for them.

## Why

As described on #3374, we currently go to great lengths to avoid calls
to `{FileHandle, FileSlice}::read_bytes`, because each of them would
acquire a lock on a Postgres `Buffer`, and then copy out of it into
`OwnedBytes`.

But `OwnedBytes` is not actually "owned": it's reference counted, and
can be backed by a structure that borrows from shared memory, like a
`Buffer`. This allows us to avoid copying `Buffer`s, and to instead
slice into them.

Additionally, because the `Buffer`s of a `BytesList` are immutable, we
do not need to hold read locks on them after loading them the first
time: a pin is sufficient.

## How

Introduces a method to assert that a `Buffer` is immutable, and release
its read lock to convert it into the `ImmutablePage` type, which can be
wrapped into `OwnedBytes`.

## Tests

Some benchmarks show up to 1.4x speedups; none show slowdowns.