ci: Split SchemaBot into two PRs to enable external contributors to use it (#3480)
# Ticket(s) Closed

- Closes #N/A

## What
Currently, external contributors cannot use SchemaBot because the
workflow is unable to post comments to the PR. To solve this, we split
the SchemaBot workflow in two:

- The same workflow runs on PRs, generates the schema diff (this worked
before) but instead of posting a comment, it uploads it as an artifact.
This is possible since it can be uploaded to the external contributor's
fork, on which they have permissions.
- We add a `workflow_run` CI job which listens for the completed
workflow, reads the artifact, and posts it to the PR. That workflow runs
from within our repository, and so is able to post a PR comment.

The key is to avoid running `actions/checkout` in the `workflow_run`
job, since it gives external contributors access to our repository.

I've also removed the `actions/checkout` using the ParadeDB Bot token,
which was preventing the workflow from running for external contributors
altogether, before even failing to post a comment.

## Why
Enable external contributors to use SchemaBot without creating security
risks.

## How
^

## Tests
This needs to be merged to `main` for the comment posting job to be
tested, but the schema generation still works.