ci: Check for existence of SQL script (#3491)
# Ticket(s) Closed

- Closes #N/A

## What
We kept forgetting SQL upgrade scripts and burning our users... This
prevents issuing a release without the upgrade script that upgrades to
that version.

## Why
So users don't face issues of "no upgrade path to version X".

## How
Check for the existence of a SQL upgrade script with the upcoming
release version in its filename.

## Tests
Missing script:
https://github.com/paradedb/paradedb/actions/runs/19057075473/job/54429579453
Found script:
https://github.com/paradedb/paradedb/actions/runs/19057104178/job/54429660239