perf: Pipeline the execution of window aggregates (#3450)
## What

Pipeline the execution of window aggregates so that they are computed in
the same query pass with TopDocs/TopN.

## Why

#3312 added support for executing aggregate queries in window functions,
but the implementation there executed aggregates _after_ executing TopN,
in each worker.

## How

Add space in the `ParallelScanState` for aggregate responses (if
aggregations are requested), and collect aggregate responses there.
After a TopN query, wait for all workers to have produced their
responses, and then finalize the aggregate.

---------

Co-authored-by: Mohammad Dashti <mdashti@gmail.com>