chore: Allow for optional MVCC filtering of window aggregates (#3494)
## What

This change adds infrastructure necessary to toggle the use of MVCC
filtering at query time. It does not yet actually add an argument to do
so.

## Why

MVCC filtering can cost anywhere between 2-10x performance for
aggregations: we will reduce this over time by more tightly integrating
with the visibility map, but in cases where the visibility map is
not/less up to date, there is not much we can do other than accessing
the heap.

## How

Introduce a wrapper type `TopNAuxiliaryCollector` which carries an
aggregation collector, and an optional visibility checker. The wrapper
is necessary because iff an aggregation will be executed, we want to
apply MVCC filtering before _both_ of the collectors, and so a
`MVCCFilterCollector` needs to be wrapped around the compound collector.