ci: Properly account for beta releases in SQL script verification (#3502)
# Ticket(s) Closed

- Closes #N/A

## What
I added the ability to check, before a release, that the SQL script
exists. But did not account for beta releases having `-rc.X` in their
name. This modifies the verification to strip the `-rc.X` tag and check
that the SQL script exists.

## Why
So it works for beta releases too.

## How
^

## Tests
^