fix: FSM freelist sparsity causes too many page reads/lock acquisitions (#3509)
# Ticket(s) Closed

- Closes #

## What

The `extend_freelist` function in V2 FSM had a "fast path splice" that
could inject tons of tiny, partially-full blocks into the freelist. If
it saw that a freelist block was full, it would blindly just assume the
next block was also full and inject a new block between the current and
next block.

However, this fails if a lot of `extend_freelist` calls come in
concurrently -- each call would inject a partially-filled block in the
middle of the list, leading to a very sparse freelist.

The solution is to peek at the next block in the freelist, and only do
the fast path if both blocks are full.

Additionally, we have seen situations in prod (likely related to the
above) where freelists contain entirely empty blocks. The existing
freelist isn't good at self-healing -- we have modified `drain` to now
remove these empty blocks from the freelist.

Finally, `drain`s call to `extend_freelist` has been moved to the end
after all blocks have been drained. This prevents multiple calls to
`extend_freelist`, which is expensive (it can do tree mutations).

## Why

## How

## Tests

See two added regression tests, both of which fail on `main`