fix: Allow up to 2 background mergers (#3519)
# Ticket(s) Closed

- Closes #

## What

Addresses two problems:

1. Right now only one background merger is allowed per index. We
actually want 2, because we don't want a long-running merge to prevent
smaller segments from being merged.
2. In production I've seen an issue where background workers were not
launching because we currently keep track of whether a merge is
happening by writing its PID to disk. I'm seeing stale PIDs being
considered live merges, or merges not being correctly recorded in the
background merger page.

## Why

## How

This PR moves us to a stateless way of tracking background merging:
pins.

The metadata page allocates 2 blocks. Whenever we want to start a
background merge, we try and take a conditional cleanup lock on one of
these pages. Then, we immediately drop the lock but keep the pin until
the merge has finished, which prevents any further cleanup locks from
being taken on this page.

The first block is pinned when the largest layer size of the merge is
under `100mb`, the second is pinned for larger layers. This guarantees
that small merges are not held back by a large one.

## Tests