ci: Fix marking Docker images as latest (#3530)
# Ticket(s) Closed

- Closes #N/A

## What
We had changed the way we check for latest to be via the GitHub Releases
tags, and since we are just past releasing the current tag, we need to
account for the version being equal.

## Why
We were no longer properly marking releases as latest.

## How
Greater than or equal.

## Tests
Locally, yes