fix: Mutable segment corruption when reading beyond number of entries (#3618)
# Ticket(s) Closed

- Closes #

## What

Under physical replication, we've observed a rare issue where entries of
the mutable segment fail to deserialize.

This always seems to happen when we are reading beyond the actual length
of the mutable segment list. For instance, the mutable segment list only
contains 400 entries, but we try and deserialize entry 401.

I don't yet have a perfect theory for why this is happening, but
stopping the reading of the merge segment list when we've reached the
number of entries seems to be working as a stopgap.

## Why

## How

## Tests