ci: Fix prebuilt binaries publish (#3637)
# Ticket(s) Closed

- Closes #N/A

## What
We recently merged a change that broke our deployments for prebuilt
extensions. This PR fixes that, and brings the deploys up to date:

- Remove deployment of RHEL 8 Debian 11, which EOL-ed last year
- Add Debian 13 and RHEL 10, and fix their deployment
- Apt-key was deprecated on Debian 13, so I switched all deployments to
use the new approach
- Updated the filtering workflow to correctly check for `-rc.` in the
git ref instead of `-`, which would incorrectly ignore deployments when
manually triggered from a branch with a hyphen in it.

## Why
Working releases!

## How
^

## Tests
Manual triggered releases which were sent to our `v0.0.0` testing tag.
They all passed.