feat: add support for PG18 (#3375)
# Ticket(s) Closed

- Closes #2723 

## What
Add support for Postgres 18 to `pg_search`. I don't yet default to PG18
for our builds, though. That can come later. We'll also want to add PG18
to the prebuilt CI binaries. This can come in a separate PR

I also upgraded pgvector to 0.8.1, which is necessary for PG18 support.

## Why
Lots of community users are waiting for this, and I had an hour to
spare.

## How
Mostly reading Postgres source and compare REL_17_STABLE and
REL_18_STABLE.

## Tests
I can compile and run the full quickstart locally, and CI should pass
for the full test suite.

---------

Signed-off-by: Philippe Noël <21990816+philippemnoel@users.noreply.github.com>
Co-authored-by: Stu Hood <stuhood@gmail.com>
Co-authored-by: Jubilee <workingjubilee@gmail.com>
Co-authored-by: Mithun Chicklore Yogendra <mithun.cy@gmail.com>
Co-authored-by: Mohammad Dashti <mdashti@gmail.com>