perf: Trigger TopN in the presence of `unnest` (#3677)
## What

Allow the `TopN` scan to be triggered in the presence of `unnest` by
retrieving the `limit` from the parse (iff we recognize all
set-returning-functions that were used).

## Why

As reported in #3622, in the presence of `unnest` `TopN` will not
trigger, and we'll instead get a `Normal` scan.

Specifically in the case of set returning functions which return more
results than are input (like `unnest`), we know that we can still apply
the TopN scan, improving performance.

## Tests

Added regress tests covering the relevant cases. #3679 will address the
other half of #3622.