fix: Fail fast for unsupported query shapes (#3679)
# Ticket(s) Closed

- Closes #3622

## What

Fail fast (with an error which points to
https://github.com/orgs/paradedb/discussions/3678) for unsupported query
shapes.

## Why

We have long debated failing fast for projected functions which we have
not been able to capture with the custom scan, but had so far decided to
continue to return an empty result. But #3622 helps to emphasize the
risks involved in failing silently, as it silently produces an empty
result.

Instead, failing fast should allow us to more rapidly learn which query
shapes matter, and avoid users being confused about whether an empty
result is intentional.