Remove WITH clause allowing multi-column indexes to be used for key_field
- find the smallest multi-row index to use a KF at build time
- at query time do the same when BM25SearchOptions is created, then lazily cache
- when we construct our LHS RHS functions make sure it's handled as a Row
- make sure Tantivy can make a TantivyValue::Array from a Row
- works for
  - sequential scan / join behavior
  - index only scan
- fails with error for
  - MLT (can fixed but I didn't want to include more code)

One caveat is that it does a lot more work each time a relation is opened, but
I can't measure any difference in practice.

We could also remove the build time checks and just emit a warning if you don't
have a key_field, because MLT is the only thing that is impacted and you can
add the index later with no issues.

open question: Should we enforce unique?