fix: Wrap the Jieba tokenizer to fix the token indices to be sequential (#3665)
# Ticket(s) Closed

- Closes #3664

## What

Wrapped the upstream `tantivy-jieba` JiebaTokenizer in order to set the
token positions to be sequential instead of set to the token start
offset in the original text.

This makes it behave like the other tokenizers.

## Why

The phrase search would return no matches for very obvious tests of
verbatim token-aligned pieces of text when using the Jieba tokenizer.
After some digging around, turns out the upstream `tantivy-jieba` sets
token positions to character offsets instead of sequential ordinals,
like the rest of the tokenizers do in ParadeDB.

As far as I can tell, then doing a phrase query, tokens are assumed to
be consecutively indexed and that's what results in a match with no
slop. For all other tokenizers this works fine as I see that they have
the `position` set with something like this:

```rust
self.token.position = self.token.position.wrapping_add(1);
```

But in
https://github.com/jiegec/tantivy-jieba/blob/master/src/lib.rs#L163 the
token stream generated has `position` set to `token.start`. I guess this
fits fine their use upstream, but fails here.

The fix was to just wrap `tantivy_jieba::JiebaTokenizer` to set the
positions to be sequential during tokenization.

## How

Wrapping JiebaTokenizer, intercepting the token stream and re-setting
the position to a simple incrementing integer.

## Tests

Built with the new code, tokenization returns the correct indices,
sequential.

---------

Co-authored-by: Lei Li <1715734693@qq.com>
Co-authored-by: Stu Hood <stuhood@gmail.com>