fix: Revert numeric pushdown for now (#3740)
## What

Temporarily revert #2978.

As demonstrated in the new test added in `anynumeric-conversion`, the
`AnyNumeric` conversion added in #2978 triggers during indexing, and can
cause valid `AnyNumeric` values which are being used directly in columns
(outside of JSON columns) to fail to match the schema.

## Why

To fix indexing of `AnyNumeric` values while preserving the numeric
pushdown, the cases which were added to `impl TryFrom<pgrx::AnyNumeric>
for TantivyValue` will likely need to be moved to JSON-specific query
parsing codepaths.

It's also possible that our strategy for indexing `numeric` columns
needs a rethink to explicitly bound it to particular bitwidths (and then
support those bitwidths in Tantivy). Finally, we should potentially
bound numeric pushdown on JSON to values which we know JSON can
represent precisely, and then let Postgres deal with JSON values between
2^53 and 2^64, for example.