chore: Cargo upgrade (#3765)
# Ticket(s) Closed

- Closes #N/A

## What
It's been a very long time since we've upgraded our deps (aside from
Tantivy), and I figured this was relevant in our polish month. Putting
this here to see if it passes CI, and whether we want to do it or not.

## Why
Up-to-date dependencies

## How
`cargo upgrade`

Note that I did not make `cargo upgrade` upgrade the deps which require
code changes, so that would need to come separately.

## Tests
Builds locally, need to test with CI