fix: Project target list values into the custom scan's virtual slot (#3768)
# Ticket(s) Closed

- Closes #3767 

## What

In the custom scan when Postgres tells us a tuple is visible/does not
need to be heap fetched, we project it directly into the result table.
However, there's an edge case where hard-coded values in the target list
get returned as null because they're not fetched from the index and
instead need to projected from the target list.

## Why

## How

## Tests