feat: Add support for the json/b subscript [] operator (#3785)
This can be used in place of `->`, the following two cases are treated
the same:

```
WHERE metadata->'color' @@@ 'white'
WHERE metadata['color'] @@@ 'white'
```

But the second shows up as a `T_SubscriptingRef` node and not a
`T_OpExpr` node.

Note that if you create an index in these ways:

```
USING BM25(id, (metadata->'color'))
USING BM25(id, (metadata['color']))
```
Then you will have to access it with that operator as well, you can't
swap. This is the same as standard Postgres expression indexes.